Changed the way to add dependencies no longer accessible.
The system property used to add dependecies had a problem adding the files to maven repository, making them unavailable for runtime.

Followed this tutorial http://blog.dub.podval.org/2010/01/maven-in-project-repository.html
Basically we add a repository that is in the path of the code, in this case the folder lib is the repository for old files, that are no longer available.




Former-commit-id: 11479f6d5f9d2731b33156c9b40cf4bf2f1588b1